before there was C++ there was C.  So far we have only been doing C, so here is the first time we are going to do the ++ part.
classes.

classes let you make your own types.  An int (integer) is a primitive type, and
you get that for free.

You have been using a class without really knowing it.
std::string is a class that handles the complexity of a string.
http://www.cplusplus.com/reference/string/string/

i.e. to get length of an int you need to call sizeof(someInt).
for a string, there is a member method on the std::string class called
length()

a lot of time was spent writing that very usefull class so that the rest of us
don't have to mess around with the dirtieness of string managment.